+++ src/OpenTelemetry.Instrumentation.AspNetCore/CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## Unreleased
 
+* Removed `netstandard2.0` and `netstandard2.1` targets
+  ([#3576](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3567))
+
 * Fixed an issue where activity started within middleware was modified by
   instrumentation library.
   ([#3498](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3498))
 
-* Updated to use Activity native support from `System.Diagnostics.DiagnosticSource`
-  to set activity status.
+* Updated to use Activity native support from
+  `System.Diagnostics.DiagnosticSource` to set activity status.
   ([#3118](https://github.com/open-telemetry/opentelemetry-dotnet/issues/3118))
   ([#3555](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3555))